wtp/webrtc/RTCPeerConnection-setRemoteDescription-offer.html

<!doctype html>
<meta charset="utf-8">
<title>RTCPeerConnection.prototype.setRemoteDescription - offer</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
  'use strict';
  
  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html
  
  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   assert_session_desc_equals()
  //   test_state_change_event()
  
  /*
    4.3.2.  Interface Definition
      [Constructor(optional RTCConfiguration configuration)]
      interface RTCPeerConnection : EventTarget {
        Promise<void>                      setRemoteDescription(
            RTCSessionDescriptionInit description);
         readonly attribute RTCSessionDescription? remoteDescription;
        readonly attribute RTCSessionDescription? currentRemoteDescription;
        readonly attribute RTCSessionDescription? pendingRemoteDescription;
        ...
      };
     4.6.2.  RTCSessionDescription Class
      dictionary RTCSessionDescriptionInit {
        required RTCSdpType type;
                 DOMString  sdp = "";
      };
     4.6.1.  RTCSdpType
      enum RTCSdpType {
        "offer",
        "pranswer",
        "answer",
        "rollback"
      };
   */
  
  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.2.3.  Otherwise, if description is set as a remote description, then run one of
              the following steps:
        - If description is of type "offer", set connection.pendingRemoteDescription
          attribute to description and signaling state to have-remote-offer.
   */
  
  promise_test(function (t) {
    var pc1 = new RTCPeerConnection();
    pc1.createDataChannel('datachannel');
  
    var pc2 = new RTCPeerConnection();
  
    test_state_change_event(t, pc2, ['have-remote-offer']);
  
    return pc1.createOffer().then(function (offer) {
      return pc2.setRemoteDescription(offer).then(function () {
        assert_equals(pc2.signalingState, 'have-remote-offer');
        assert_session_desc_equals(pc2.remoteDescription, offer);
        assert_session_desc_equals(pc2.pendingRemoteDescription, offer);
        assert_equals(pc2.currentRemoteDescription, null);
      });
    });
  }, 'setRemoteDescription with valid offer should succeed');
  
  promise_test(function (t) {
    var pc1 = new RTCPeerConnection();
    pc1.createDataChannel('datachannel');
  
    var pc2 = new RTCPeerConnection();
  
    // have-remote-offer event should only fire once
    test_state_change_event(t, pc2, ['have-remote-offer']);
  
    return pc1.createOffer().then(function (offer) {
      return pc2.setRemoteDescription(offer).then(function () {
        return pc2.setRemoteDescription(offer);
      }).then(function () {
        assert_equals(pc2.signalingState, 'have-remote-offer');
        assert_session_desc_equals(pc2.remoteDescription, offer);
        assert_session_desc_equals(pc2.pendingRemoteDescription, offer);
        assert_equals(pc2.currentRemoteDescription, null);
      });
    });
  }, 'setRemoteDescription multiple times should succeed');
  
  promise_test(function (t) {
    var pc1 = new RTCPeerConnection();
    pc1.createDataChannel('datachannel');
  
    var pc2 = new RTCPeerConnection();
  
    // have-remote-offer event should only fire once
    test_state_change_event(t, pc2, ['have-remote-offer']);
  
    return pc1.createOffer().then(function (offer1) {
      return pc1.setLocalDescription(offer1).then(function () {
        return pc1.createOffer({ offerToReceiveAudio: true }).then(function (offer2) {
          assert_session_desc_not_equals(offer1, offer2);
  
          return pc2.setRemoteDescription(offer1).then(function () {
            return pc2.setRemoteDescription(offer2);
          }).then(function () {
            assert_equals(pc2.signalingState, 'have-remote-offer');
            assert_session_desc_equals(pc2.remoteDescription, offer2);
            assert_session_desc_equals(pc2.pendingRemoteDescription, offer2);
            assert_equals(pc2.currentRemoteDescription, null);
          });
        });
      });
    });
  }, 'setRemoteDescription multiple times with different offer should succeed');
  
  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.1.4.  If the content of description is not valid SDP syntax, then reject p with
              an RTCError (with errorDetail set to "sdp-syntax-error" and the
              sdpLineNumber attribute set to the line number in the SDP where the syntax
              error was detected) and abort these steps.
   */
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
  
    return pc.setRemoteDescription({
      type: 'offer',
      sdp: 'Invalid SDP'
    }).then(function () {
      assert_unreached('Expect promise to be rejected');
    }, function (err) {
      assert_equals(err.errorDetail, 'sdp-syntax-error', 'Expect error detail field to set to sdp-syntax-error');
  
      assert_true(err instanceof RTCError, 'Expect err to be instance of RTCError');
    });
  }, 'setRemoteDescription(offer) with invalid SDP should reject with RTCError');
  
  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.1.3.  If the description's type is invalid for the current signaling state of
              connection, then reject p with a newly created InvalidStateError and abort
              these steps.
     [JSEP]
      5.6.  If the type is "offer", the PeerConnection state MUST be either "stable" or
            "have-remote-offer".
   */
  promise_test(function (t) {
    var pc = new RTCPeerConnection();
    return pc.createOffer().then(function (offer) {
      return pc.setLocalDescription(offer).then(function () {
        return promise_rejects(t, 'InvalidStateError', pc.setRemoteDescription(offer));
      });
    });
  }, 'setRemoteDescription(offer) from have-local-offer state should reject with InvalidStateError');

</script>
